Fix: Not working after reopen
Steps to reproduce:
1. Create a simple app that has a button to open the qr reader
2. Run the app on Android
3. Click the button and see the camera opens
4. Use the back button to close the app (NOT FORCE CLOSE)
5. Reopen the app by clicking the launcher icon
6. Click on the button again to open the qr reader -> Nothing happens (console says that no plugin is registered on channel `qrcodereader`).

The reason for that error is:
The plugins are cleaned up after the app is closed. But the instance you keep is not cleaned up. So when the app opens again, it tries to register all plugins. But because you still have the instance, you don't do anything. So the plugin does not get registered.

With the fix it will work correctly (tested it locally).

Would be nice if this gets merged in soon :). Thx!